feat: use double for weight and counter values

+++ Haproxy.AgentCheck/Config/WeightRule.cs
@@ -5,8 +5,8 @@ namespace Lucca.Infra.Haproxy.AgentCheck.Config;
 internal class WeightRule
 {
     public SystemResponse SystemResponse { get; set; }
-    public int MinValue { get; set; } = 0;
-    public int MaxValue { get; set; } = 100;
-    public int MinWeight { get; set; } = 0;
-    public int MaxWeight { get; set; } = 100;
+    public double MinValue { get; set; } = 0;
+    public double MaxValue { get; set; } = 100;
+    public double MinWeight { get; set; } = 0;
+    public double MaxWeight { get; set; } = 100;
 }
